+++ docs/changelog.rst
@@ -31,6 +31,7 @@ CHANGELOG
 - Refonte du style icones de patrimonialité, protection et menace. Les 3 icones sont maintenant surcouchables
 - Possibilité d'ajouter des labels sur la sidebar (#729 @juggler31)
 - Ajout du paramètre `LIMIT_POINT_MAILLE` qui permet de définir le seuil à partir duquel on affiche les données en mode maille sur un atlas en mode point (défaut 500 observations)
+- Le paramètre `drop_apps_db` du fichier settings.ini est déprécié. Il est remplacer par `ATLAS_DROP_SCHEMA` qui ne supprime que le schéma atlas
 A VERIFIER :
 
 ``navbar.html``